[ACS-5645] unit test issue fix
