+++ projects/aca-content/src/lib/components/details/details.component.spec.ts
@@ -192,7 +192,7 @@ describe('DetailsComponent', () => {
 
   it('should return the icon when getNodeIcon is called', () => {
     const expectedIcon = 'assets/images/ft_ic_folder';
-    spyOn(component['nodeActionsService'], 'getNodeIcon').and.returnValue(expectedIcon);
+    spyOn(contentApiService, 'getNodeIcon').and.returnValue(expectedIcon);
     fixture.detectChanges();
     const result = component.getNodeIcon(mockNode);
     expect(result).toContain(expectedIcon);
@@ -200,18 +200,10 @@ describe('DetailsComponent', () => {
 
   it('should return the icon when getNodeIcon is called', () => {
     const expectedIcon = 'assets/images/ft_ic_folder';
-    spyOn(component['nodeActionsService'], 'getNodeIcon').and.returnValue(expectedIcon);
+    spyOn(contentApiService, 'getNodeIcon').and.returnValue(expectedIcon);
     fixture.detectChanges();
     const result = component.getNodeIcon(mockNode);
     expect(result).toContain(expectedIcon);
-  });
-
-  it('should subscribe to store and update isNodeLocked', () => {
-    const mockSelection = { file: { entry: { name: 'test', properties: {}, isLocked: false } } };
-    spyOn(store, 'select').and.returnValue(of(mockSelection));
-    fixture.detectChanges();
-    expect(store.select).toHaveBeenCalled();
-    expect(component.isNodeLocked).toBe(false);
   });
 
   it('should unsubscribe from observables on component destroy', () => {
