+++ projects/aca-shared/src/lib/components/info-drawer/info-drawer.component.spec.ts
@@ -32,7 +32,6 @@ import { InfoDrawerComponent } from './info-drawer.component';
 import { LibTestingModule } from '../../testing/lib-testing-module';
 import { AppExtensionService } from '../../services/app.extension.service';
 import { ContentApiService } from '../../services/content-api.service';
-import { MatDialogModule } from '@angular/material/dialog';
 
 describe('InfoDrawerComponent', () => {
   let fixture: ComponentFixture<InfoDrawerComponent>;
@@ -92,7 +91,7 @@ describe('InfoDrawerComponent', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [LibTestingModule, InfoDrawerComponent, MatDialogModule],
+      imports: [LibTestingModule, InfoDrawerComponent],
       providers: [
         { provide: AppExtensionService, useValue: extensionServiceMock },
         { provide: Store, useValue: storeMock }
@@ -218,7 +217,7 @@ describe('InfoDrawerComponent', () => {
 
   it('should return the icon when getNodeIcon is called', () => {
     const expectedIcon = 'assets/images/ft_ic_folder';
-    spyOn(component['nodeActionsService'], 'getNodeIcon').and.returnValue(expectedIcon);
+    spyOn(contentApiService, 'getNodeIcon').and.returnValue(expectedIcon);
     fixture.detectChanges();
     const result = component.getNodeIcon(mockNode);
     expect(result).toContain(expectedIcon);
